requirements.txt
fastapi==0.111.0
asgiref==3.4.1
pydantic==2.7.3
uvicorn==0.30.1
starlette==0.37.2
python-decouple==3.8
requests~=2.28.1
pytest-asyncio==0.18.3
SQLAlchemy==1.4.32
psycopg2==2.9.6
psycopg2-binary==2.9.6